[refactor] NJT-53 NoteCardの肥大化抑制のため分割リファクタリング
